+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - docs:update Polaris test environment ip.
 - fix:fix custom fallback exception.
 - feat:sct-all package is now available as a shaded uber-jar.
+- fix:use path parameter in `@FeignClient` for circuit-breaker.